Extracted getBaseUrl function to a utils file for better reusability
